Embedded Types
The following commit is in part a revert from an earlier commit.
In order to gain access to fields from embedded structs we need
to expand them correctly. To also be backwards compatible with the
previous commit, we also show the embedded nature of that struct.

Eventually we should turn that later off, but considering we've
begun actively developing using this for pylibjuju we should
remain as backwards compatible as possible.